Fixing api selectors and continue removal of statee sections
