+++ src/app/features/applications/application/application-base.component.spec.ts
@@ -1,4 +1,3 @@
-import { appReducers } from '../../../store/reducers.module';
 import { it } from '@angular/cli/lib/ast-tools/spec-utils';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
@@ -8,16 +7,14 @@ import { StoreModule } from '@ngrx/store';
 import { CoreModule } from '../../../core/core.module';
 import { MDAppModule } from '../../../core/md.module';
 import { SharedModule } from '../../../shared/shared.module';
-import { paginationReducer } from '../../../store/reducers/pagination-reducer/pagination.reducer';
+import { appReducers } from '../../../store/reducers.module';
 import { getInitialTestStoreState } from '../../../test-framework/store-test-helper';
 import { ApplicationBaseComponent } from './application-base.component';
 import { ApplicationEnvVarsService } from './build-tab/application-env-vars.service';
 import {
   ApplicationStateIconComponent,
 } from './build-tab/application-state/application-state-icon/application-state-icon.component';
-import {
-  ApplicationStateIconPipe,
-} from './build-tab/application-state/application-state-icon/application-state-icon.pipe';
+import { ApplicationStateIconPipe } from './build-tab/application-state/application-state-icon/application-state-icon.pipe';
 import { ApplicationStateComponent } from './build-tab/application-state/application-state.component';
 import { ApplicationStateService } from './build-tab/application-state/application-state.service';
