+++ src/app/store/reducers.module.ts
@@ -1,30 +1,21 @@
-import { requestDataReducer, requestReducer } from './reducers/api-request-reducers.generator';
-import { CNSISEffect } from './effects/cnsis.effects';
-import { EndpointSchema } from './actions/cnsis.actions';
-import { ApiActionTypes, RequestTypes } from './actions/request.actions';
-import { combineReducers, StoreModule, ActionReducerMap, State } from '@ngrx/store';
+import { NgModule } from '@angular/core';
+import { ActionReducer, StoreModule } from '@ngrx/store';
+import { StoreDevtoolsModule } from '@ngrx/store-devtools';
 import { storeFreeze } from 'ngrx-store-freeze';
 import { storeLogger } from 'ngrx-store-logger';
 
 import { environment } from '../../environments/environment';
-import { AppState } from './app-state';
-import { appMetadataRequestReducer } from './reducers/app-metadata-request.reducer';
-import { appMetadataReducer } from './reducers/app-metadata.reducer';
+import { actionHistoryReducer } from './reducers/action-history-reducer';
+import { requestDataReducer, requestReducer } from './reducers/api-request-reducers.generator';
 import { authReducer } from './reducers/auth.reducer';
 import { cnsisReducer } from './reducers/cnsis.reducer';
 import { createAppReducer } from './reducers/create-application.reducer';
 import { dashboardReducer } from './reducers/dashboard-reducer';
-import { createPaginationReducer } from './reducers/pagination-reducer/pagination.reducer';
-import { uaaSetupReducer } from './reducers/uaa-setup.reducers';
-import { NgModule } from '@angular/core';
-import { StoreDevtoolsModule } from '@ngrx/store-devtools';
-import { routerReducer } from '@ngrx/router-store';
-import { RouterStateSnapshot, Params } from '@angular/router';
-import { actionHistoryReducer } from './reducers/action-history-reducer';
-import { MetadataState } from './types/app-metadata.types';
 import { listReducer } from './reducers/list.reducer';
-import { generateDefaultState } from './reducers/api-request-reducer/request-helpers';
-import { OtherEntityStateNames } from './types/other-entity.types';
+import { generateRequestPaginationReducer } from './reducers/pagination-reducer.generator';
+import { appMetaDataReducer } from './reducers/pagination-reducer/app-metadata-reducer.generator';
+import { uaaSetupReducer } from './reducers/uaa-setup.reducers';
+import { ActionReducerMap } from '@ngrx/store/src/models';
 
 
 export function logger(reducer): any {
@@ -32,38 +23,19 @@ export function logger(reducer): any {
   return storeLogger()(reducer);
 }
 
-const appMetadataReducers: ActionReducerMap<MetadataState> = {
-  values: appMetadataReducer,
-  requests: appMetadataRequestReducer
-};
-
-export function appMetaDataReducer(state, action): MetadataState {
-  // https://github.com/ngrx/platform/issues/116#issuecomment-317297642
-  return combineReducers<MetadataState>(appMetadataReducers)(state, action);
-}
-
-export function paginationReducer() {
-  return createPaginationReducer([
-    RequestTypes.START,
-    RequestTypes.SUCCESS,
-    RequestTypes.FAILED
-  ]);
-}
-
 export const appReducers = {
   auth: authReducer,
   uaaSetup: uaaSetupReducer,
   cnsis: cnsisReducer,
-  pagination: paginationReducer(),
+  pagination: generateRequestPaginationReducer(),
   request: requestReducer(),
   requestData: requestDataReducer(),
   dashboard: dashboardReducer,
   createApplication: createAppReducer,
-  appMetadata: appMetaDataReducer,
+  appMetadata: appMetaDataReducer(),
   actionHistory: actionHistoryReducer,
-  lists: listReducer,
-  // routerReducer: routerReducer,  // Create action for router navigation
-};
+  lists: listReducer
+} as ActionReducerMap<{}>;
 
 let metaReducers = [];
 if (!environment.production) {
